requirements.txt
feedparser==6.0.11
python-dateutil==2.9.0
google-genai==1.2.0
Jinja2==3.1.4
